+++ rules/aggregate_proj.py
@@ -1,12 +1,12 @@
 rule aggregate_proj:
     input:
-        "results/projection/raw/{project}_{activity}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}"
+        "results/projection/raw/{project}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}"
     output:
-        "results/projection/means/{area}_{project}_{activity}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}_{aggregation}_{period}.nc"
+        "results/projection/means/{area}_{project}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}_{aggregation}_{period}.nc"
     log:
-        "results/logs/aggregate_proj_{area}_{project}_{activity}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}_{aggregation}_{period}.log"
+        "results/logs/aggregate_proj_{area}_{project}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}_{aggregation}_{period}.log"
     benchmark:
-        "results/benchmarks/aggregate_proj_{area}_{project}_{activity}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}_{aggregation}_{period}.benchmark.txt"
+        "results/benchmarks/aggregate_proj_{area}_{project}_{domain}_{institute}_{model}_{experiment}_{ensemble}_{rcm}_{downscaling}_{base}_{aggregation}_{period}.benchmark.txt"
     threads: 1
     resources:
         mem_mb=1000